Yahoo!  Fully functional remote control
The remote control is working just fine.  It is dependent
on two files.  The williams/joystick loaded onto the
Esplora and williams_joystick_control loaded onto the bot.

There also seems to be a strange power problem where the
servos jerk for no apparaent reason.  When I watch the
serial stream of signals from the controller they are stable.

I even filtered the commands a little to take out sudden
peturbations, but the jerk is still present in teh bot.  I think
it likely has something to do with power because williams flashes
his smile light occassionally and I've got nothing programmed
for that pin right now.